Fix performance issue, leaving debug code intact
The issue arose with the PHP function uniqid, which uses randomness
to create 'unique' IDs...  I believe it might have been using a
blocking source for randomness, as could take over 8 seconds to get
and ID sometimes.  Anyway...  Created a function which uses serial
numbers, which seems to have solved it.